Create PasswordV4.py
This Python program checks if a user’s password is strong enough according to its established standards. It guarantees that the password has at least eight characters, and it has a combination of upper case, lower case, numeric digits and special characters. First, it invites the user to input a password and then goes through each letter in order to check if it meets these requirements. If all criteria are met by the password, then it's considered as being strong and valid whereas in case they fail this test they are invalid. Also, after every checking whether to validate another one or not can be made for several passwords within one session. This feature makes it quite useful for people who want their passwords to meet the standard security levels.